+++ src/Fable.AST/RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 3.0.0-nagareyama-beta-001
+
+* Beta release
+* Use string literals in ImportInfo
+
 ### 3.0.0-nagareyama-alpha-003
 
 * Add optional string tag to TypeCast